Module fenix_rooms::api::handlers

Handlers for the REST API

Each handler reads the request if need be, gets the information from getters and returns a Response accordingly,

Functions

check_in_handler
Checks in in the Database

check_out_handler
Checks out in the Database

create_room_handler
Creates a Room in the Database

create_user_handler
Creates a User in the Database

id_handler
Handler for IDs using the FenixEDU API. The id sent in the url will be processed.

path_handler
Translate the ID's to names

rooms_handler
Gets the list of rooms in the Database

spaces_handler
Handler for the top level spaces at IST