Function fenix_rooms::api::handlers::create_room_handler

pub fn create_room_handler(request: &mut Request) -> PencilResult

Creates a Room in the Database

Create a room in the database with the specified fenix_id, capacity, location and user_id in the body. Only the admin can create rooms in the DB.

Arguments

request - The request sent by the client

Output

A Response with a JSON messsage and correct status code.