Function fenix_rooms::api::handlers::path_handler

pub fn path_handler(request: &mut Request) -> PencilResult

Translate the ID's to names

This handler translates id's to names letting you browse the spaces API hierarchically. Previously /api/id/<id> and now /api/path/level1/level2/level3. Keep in mind that by increasing the amount of levels in the path the more GET requests are made. The response time is now dependent on the responsiveness of the FenixEDU API. Also this might cause a mini DDOS.

Output

JSON message with the contents of the requested space.