Function fenix_rooms::api::handlers::spaces_handler

pub fn spaces_handler(_: &mut Request) -> PencilResult

Handler for the top level spaces at IST

The handler calls utils::get_spaces_from_id() to perform the GET request required. If the request was successful its contents will be sent as JSON. Otherwise an error will be sent, provided by the function.

Arguments

_ => The Pencil framework requires the handler signature to be of the type fn(&mut Request) -> PencilResult. However, this handler doesn't require access to the Request sent by the client. So we declare as unused with _.

Return Value

Error if the utils::get_spaces_from_id() fails. Otherwise read the contents and send it as JSON.